Disable ConnPool overflow by default
Oveflow should only be enabled when the number of active threads
using the pool is not known or bounded and there is an
occasional need for a few extra connections.

By default, users who have mistakenly set pool_size too low but
are not aware of overflow behavior will pay a performance penalty
from an increase in the number of opened and closed connections.
To prevent this, overflow is now disabled by default.