Add conns to pool more readily when prefill=False
Before this change, if the ConnectionPool was created with
prefill=False, connections would only be added to the pool
when there was concurrent demand for connections.

After this change, if prefill=False and pool_size=N, the
first N operations will each result in a new connection
being added to the pool.