Refactor out AbstractPool, make attrs accessible
Because ConnectionPool was the only class that extended
AbstractPool, and extending this class with another would
be messy anyhow, the class AbstractPool has been refactored
out.

In addition, several attributes of ConnectionPool have been
moved out of __init__ as kwargs into alterable attributes.
If used as kwargs, these will continue to work, but they may
now be altered after construction.