Convert packing system to pass functions
Instead running through the logic of what type
a piece of data each time that it's packed or
unpacked, we can determine the correct basic function
to use for packing and unpacked one time on initialization
and just use that.

This will be needed to make the packing and unpacking of
CompositeTypes efficient, but has a side effect of being
more efficient for normal types as well.