Introduce SASL transport support
While the TSaslClientTransport is not in mainline thrift,
it will be bundled in the connection module.

The user may now provide a transport_factory function that
will be used to create transports for new connections. As a
convenience, the pycassa.connection.make_sasl_transport_factor()
function will capture SASL parameters in a closure and return
a factory function, but the user is also free to define their
own factory function.